pump.fun/requirements.txt
tweepy==4.14.0
solana==0.30.2
solders==0.18.1
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.11.2
base58==2.1.1
rich==13.9.4
beautifulsoup4==4.12.2
click==8.1.7
selenium==4.16.0
webdriver_manager==4.0.1
textblob==0.17.1
networkx==3.2.1
emoji==2.8.0
async_lru==2.0.4
oauthlib==3.2.2
tweety-ns==1.0.2
httpx[http2]==0.25.2
chromium-browser==120.0.6099.109
playwright==1.41.2